pyproject.toml
[project]
name = "data-engineering-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-openlineage>=1.12.2",
    "apache-airflow>=2.10.2",
    "cassandra-driver>=3.29.2",
    "kafka-python-ng>=2.2.3",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "pyspark>=3.5.3",
    "spark>=0.2.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
]
